moved login to functions.js file
