+++ src/functions/functions.js
@@ -52,6 +52,22 @@ export const alreadyUser = async (user) => {
 		console.log('couldnt check if user')
 	}
 }
+export const login = async () => {
+	const result = await auth.signInWithPopup(googleProvider)
+	let adminCheck = await checkAdminEmail(result.user.email)
+	// console.log(adminCheck)
+	let restaurantCheck = await checkRestaurantEmail(result.user.email)
+	// console.log(restaurantCheck)
+	let user = result.user
+	let jsonifiedUser = JSON.parse(JSON.stringify(user))
+	console.log(jsonifiedUser)
+	let userCheck = await alreadyUser(user)
+	console.log(userCheck)
+	if (!userCheck) { postUser(jsonifiedUser) }
+	adminCheck ? result.user.isAdmin = true : restaurantCheck ? result.user.isRestaurant = true : user = user
+	console.log('this is the user', user)
+	props.updateUser(user)
+}
 export const postUser = async (user) => {
 	let ref = firebase.database().ref('users')
 	let username = user.displayName
